By default gutter-grid will NOT support legacy browsers
It is still possible to support them but you now need to set $grid-legacy-support to "true" if you want the extra css needed for them to render.

This legacy css adds a lot of extra bloat to the output css if left in and most users don't need to support these crippled old browsers any more so it's adding extra bloat for nothing.