$align defaults to left; $wrap defaults to true (Mixin)
CSS Grid is going to completely kill off any need for Gutter Grid soon.

IE 11 is essentially the only reason Gutter Grid is still useful right now.

With IE 8 and 9 essentially dead, it doesn't make sense to hamper the experience of IE11 developers for the sake of these dead browsers.

I'm optimising the code to make the experience as enjoyable as possible for IE11 devs. Unfortunately it means the defaults will be different depending on if the legacy setting is turned on or not.

If the legacy setting is turned on then they get the same defaults as version 3.0.0.