Creating a main.legacy.scss file
If a user opens up the site in IE 8 or 9, I still want the site to work since I boast IE8/9 support.

I want to show users on modern browsers what the default settings output though.

So in IE8 & 9, they are going to load main.legacy.css with the legacy setting turned on. All other browsers are going to load the regular main.css file.