Adding calc() and columns count breakpoints support
I've added a new setting called $grid-calc-support. It is set to whatever the opposite of $grid-legacy-support is by default.

When calc support is turned on, it will output calc(100% / x) where x is the number of columns instead of pre-calculating all of the percentages for you.

Custom column breakpoint settings can now also be set by defining a number of columns rather than specifying a % width.