BREAKING CHANGE: $grid-break-points setting
In version 3, the $grid-break-points setting would base the number of columns on the index of a set of breakpoint values.

This made it difficult to tell what the number of columns for a particular set of breakpoints was for.

This update now allows you to explicitly state the default column count for each set of breakpoints.