Adding grid--wrap support implicitly by default to grid--align-stretch class
The goal is to get IE8 & 9 looking as close to the modern browsers as possible.
I can't really do that on the align properties without also allowing wrapping at the same time.

In order to make grid--align-stretch consistent with the other align properties, I'm making it wrap by default as well