[breaking change] grid__wrapper class altered & grid__paddedWrapper class removed
The grid__wrapper class now adds 0.1px of bottom padding rather than overflow hidden.

0.1px of padding gets rounded down to 0px in browsers but it still prevents margin collapse in all browsers! (Thanks to
Lu Nelson for the tip!)

This means that the `grid__wrapper` class can be used to contain boxes with box shadows without having to worry about an ugly 1px of bottom padding.

With the `grid__wrapper` class now able to safely contain box-shadows, the grid__paddedWrapper class has been made obsolete.

grid__paddedWrapper hasn't been in Gutter Grid for very long so I'm just going to remove it rather than depricate it.

The mixin system is unafected by this change since the wrapper is not controlled by the mixin. it is controlled manually by the user.